MEDO-563: updated algorithm on evaluation clauses
